+++ src/GitAttributes/Writer.php
@@ -116,7 +116,9 @@ final readonly class Writer implements WriterInterface
             ];
         }
 
-        if ([] !== $rows && 'raw' === $rows[array_key_last($rows)]['type'] && '' === $rows[array_key_last($rows)]['line']) {
+        if ([] !== $rows && 'raw' === $rows[array_key_last($rows)]['type'] && '' === $rows[array_key_last(
+            $rows
+        )]['line']) {
             array_pop($rows);
         }